增加商品模块集成 运费bug修正 根据网关去除payment前缀

+++ freight/src/main/java/com/example/freight/model/bo/FreightModelDetail.java
@@ -21,5 +21,5 @@ public interface FreightModelDetail {
     * @Author: alex101
     * @Date: 2020/12/14
     */
-    Long getFreight(int weight, Long count);
+    Long getFreight(Long weight, Long count);
 }
